+++ src/index.js
@@ -1,2 +1,3 @@
 export * from './format'
 export {default as numberformat} from './format'
+export * from './parse'